added seeds for tags and mechants

+++ models/transaction.rb
@@ -8,7 +8,7 @@ class Transaction
     @id = options['id'].to_i if options['id']
     @merchant_id = options['merchant_id'].to_i
     @tag_id = options['tag_id'].to_i
-    @amount = options['amount'].to_i
+    @amount = options['amount'].to_f
   end
 
   def save()
